+++ src/HW01/View.java
@@ -1,0 +1,17 @@
+package HW01;
+public class View {
+    public void showMenu(){
+            System.out.println("выберите пункт меню \n");
+            System.out.println("1 - загрузить заранее подготовленый тестовый список имен");
+            System.out.println("2 - продолжить с сохраненного списка");
+            System.out.println("3 - ввести данные о новом человеке");
+            System.out.println("4 - изменить данные человека");
+            System.out.println("5 - ввести данные об окруженнии человека");
+            System.out.println("6 - вывести всех всех людей в списке");
+            System.out.println("7 - введите имя человека для поиска");
+            System.out.println("8 - сохранить список");
+            System.out.println("9 - сортировка списка по имени");
+            System.out.println("10 - сортировка списка по дате рождения");
+            System.out.println("0 - завершить работу");
+    }
+}
