+++ src/HW01/model/Basic.java
@@ -1,0 +1,30 @@
+package HW01.model;
+
+public class Basic {
+    public String getName() {
+        return null;
+    }
+
+    public void addChild(String string) {
+    }
+
+    public void setDateOfBirth(Integer birthDate) {
+    }
+
+    public void setDateOfDeath(Integer deathDate) {
+    }
+
+    public void setFather(String fatherName) {
+    }
+
+    public void setMother(String motherName) {
+    }
+
+    public Integer getDateOfBirth() {
+        return null;
+    }
+
+    public Integer getDateOfDeath() {
+        return null;
+    }
+}
